feat: Rebuilt CLI setup
This is a complete revamp of how monetr's CLI code is structured, it
should make it easier to maintain commands going forward as well as make
it so that commands can actually be tested where they were very
difficult to test before this.
