+++ server/storage/storage.go
@@ -7,9 +7,15 @@ import (
 	"path"
 	"strings"
 
+	"github.com/aws/aws-sdk-go/aws"
+	"github.com/aws/aws-sdk-go/aws/credentials"
+	"github.com/aws/aws-sdk-go/aws/session"
+	"github.com/aws/aws-sdk-go/service/s3"
 	"github.com/google/uuid"
+	"github.com/monetr/monetr/server/config"
 	"github.com/monetr/monetr/server/models"
 	"github.com/pkg/errors"
+	"github.com/sirupsen/logrus"
 )
 
 //go:generate go run go.uber.org/mock/mockgen@v0.5.2 -source=storage.go -package=mockgen -destination=../internal/mockgen/storage.go Storage
@@ -100,4 +106,65 @@ func getContentTypeByPath(filePath string) (ContentType, error) {
 func GetContentTypeIsValid(contentType string) bool {
 	_, ok := contentTypeExtensions[ContentType(contentType)]
 	return ok
+}
+
+// GetStorage is meant to be called by the application initially starting up, it
+// is simply provided a log entry and a configuration. It will the construct the
+// appropriate storage driver implementation and return it to the caller. If an
+// invalid configuration is provided then this will return an error. If storage
+// is not configured then this will return nil.
+func GetStorage(
+	log *logrus.Entry,
+	configuration config.Configuration,
+) (fileStorage Storage, err error) {
+	if !configuration.Storage.Enabled {
+		log.Trace("file storage is not enabled")
+		return nil, nil
+	}
+
+	switch configuration.Storage.Provider {
+	case "s3":
+		log.Trace("setting up file storage interface using S3 protocol")
+		s3Config := configuration.Storage.S3
+		awsConfig := aws.NewConfig().WithS3ForcePathStyle(s3Config.ForcePathStyle)
+		if endpoint := s3Config.Endpoint; endpoint != nil {
+			awsConfig = awsConfig.WithEndpoint(*endpoint)
+		}
+
+		if useEnvCredentials := s3Config.UseEnvCredentials; useEnvCredentials {
+			awsConfig = awsConfig.WithCredentials(credentials.NewEnvCredentials())
+		} else if s3Config.AccessKey != nil {
+			awsConfig = awsConfig.WithCredentials(credentials.NewStaticCredentials(
+				*s3Config.AccessKey,
+				*s3Config.SecretKey,
+				"", // Not requiured since we aren't using temporary credentials.
+			))
+		}
+
+		if s3Config.Region != "" {
+			awsConfig = awsConfig.WithRegion(s3Config.Region)
+		}
+
+		session, err := session.NewSession(awsConfig)
+		if err != nil {
+			return nil, errors.Wrap(err, "failed to create s3 session")
+		}
+
+		client := s3.New(session)
+
+		fileStorage = NewS3StorageBackend(log, s3Config.Bucket, client)
+	case "filesystem":
+		log.Trace("setting up file storage interface using local filesystem")
+		fileStorage, err = NewFilesystemStorage(
+			log,
+			configuration.Storage.Filesystem.BasePath,
+		)
+	default:
+		return nil, errors.Errorf(
+			"invalid storage provider: %s",
+			configuration.Storage.Provider,
+		)
+	}
+
+	return fileStorage, err
 }
